+++ ui/index.html
@@ -6,11 +6,11 @@
     <meta name="description"
           content="wiretap is a tool that sniffs your API traffic and validates it against an OpenAPI specification.">
     <script>
-        //localStorage.setItem('wiretapPort', '%WIRETAP_PORT%')
+        localStorage.setItem('wiretapPort', '%WIRETAP_PORT%')
         localStorage.setItem('wiretapTLS', '%WIRETAP_TLS%')
-        //localStorage.setItem('wiretapHost', '%WIRETAP_HOST%')
-        localStorage.setItem('wiretapHost', 'localhost')
-        localStorage.setItem('wiretapPort', '9092');
+        localStorage.setItem('wiretapHost', '%WIRETAP_HOST%')
+        //localStorage.setItem('wiretapHost', 'localhost')
+        //localStorage.setItem('wiretapPort', '9092');
         //localStorage.setItem('wiretapTLS', 'true');
         const version = '%WIRETAP_VERSION%';
         const existingVersion = localStorage.getItem('wiretapVersion');
