+++ ui/src/index.ts
@@ -64,5 +64,5 @@ import './wiretap';
 
 // Set the base path to the folder you copied Shoelace's assets to
 import {setBasePath} from '@shoelace-style/shoelace/dist/utilities/base-path.js';
-setBasePath('/shoelace');
-//setBasePath('/assets/shoelace');
+//setBasePath('/shoelace');
+setBasePath('/assets/shoelace');
